refactor(game-engine): extract handleBlitz (S27.8.8)
Extracts the BLITZ handler (~228 lines) from actions.ts into
actions/blitz-handler.ts. No behavior change — pure structural
extraction to keep reducing the monolithic dispatcher towards the
S27.8 DoD target (actions.ts <= 600 lines).

After extraction: actions.ts 1517 -> 1289 lines (-15% this slice,
cumul -52.4% from origin 2706). Cleans up dead imports left over
from earlier slices: incrementTeamBlitzCount, rollBlockDiceManyWithRolls,
plus the bulk import from utils/dice that now only needs
blockResultFromRoll.

4 smoke tests (gates only — attacker/target absence, wrong team turn,
same-team target). Block/dodge/shadowing/break-tackle/armor mechanics
remain covered by mechanics/blitz.test.ts and friends. Coverage
exclusion added in vitest.config.ts (same v8 source-map quirk as
prior S27.8.x extractions).
